+++ frontend/src/components/actionInfoCard/DynamicInfoPanel.styled.tsx
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+  display: flex;
+  flex: 1;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  gap: 12px;
+  margin-bottom: 24px;
+`;
+
+export const TextWrapper = styled.div`
+  display: flex;
+  gap: 24px;
+  flex-direction: column;
+  align-items: center;
+`;
