+++ frontend/src/vault/keygen/shared/peerDiscovery/KeygenPeerDiscoveryStep/KegenPeerDiscoveryStep.styled.tsx
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+export const Content = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 40px;
+`;
+
+export const Wrapper = styled.div`
+  flex: 1;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 40px;
+`;
